refactor(sdp): introduce sdp attribute extension trait
