+++ formats/rtp/Cargo.toml
@@ -22,6 +22,7 @@ futures = "0.3.31"
 unified-io = { path = "../../unifiedio" }
 base64 = "0.22.1"
 scopeguard = "1.1"
+sdp-formats = { path = "../../formats/sdp" }
 
 [lints.clippy]
 uninlined_format_args = "allow"
