+++ formats/rtp/src/codec/mpeg4_generic/au_header/writer.rs
@@ -1,12 +1,9 @@
-use std::io;
-
+use super::{AuHeader, AuHeaderSection};
+use crate::codec::mpeg4_generic::{errors::RtpMpeg4Error, parameters::rfc3640::RtpMpeg4Fmtp};
 use bitstream_io::{BigEndian, BitWrite, BitWriter};
 use num::ToPrimitive;
+use std::io;
 use utils::traits::writer::{BitwiseWriteTo, WriteTo};
-
-use crate::codec::mpeg4_generic::{errors::RtpMpeg4Error, parameters::RtpMpeg4Fmtp};
-
-use super::{AuHeader, AuHeaderSection};
 
 pub struct AuHeaderWriteWrapper<'a>(pub &'a RtpMpeg4Fmtp, pub bool, pub &'a AuHeader);
 impl<'a, W: BitWrite> BitwiseWriteTo<W> for AuHeaderWriteWrapper<'a> {
