+++ formats/rtp/src/codec/mpeg4_generic/packet/sequencer/mod.rs
@@ -1,3 +1,14 @@
+use super::RtpMpeg4GenericPacket;
+use crate::{
+    codec::mpeg4_generic::{
+        access_unit::{AccessUnit, AccessUnitFragment},
+        errors::{RtpMpeg4Error, RtpMpeg4Result},
+        parameters::rfc3640::RtpMpeg4Fmtp,
+    },
+    errors::RtpError,
+    header::RtpHeader,
+    packet::sequencer::{RtpBufferAudioItem, RtpBufferItem, RtpBufferedSequencer},
+};
 use de_interleaving::RtpMpeg4GenericDeInterleavingBuffer;
 use fragments::RtpMpeg4GenericFragmentationBuffer;
 use tokio_util::{bytes::Buf, either::Either};
@@ -5,19 +16,6 @@ use utils::traits::{
     buffer::{GenericFragmentComposer, GenericSequencer},
     reader::ReadRemainingFrom,
 };
-
-use crate::{
-    codec::mpeg4_generic::{
-        access_unit::{AccessUnit, AccessUnitFragment},
-        errors::{RtpMpeg4Error, RtpMpeg4Result},
-        parameters::RtpMpeg4Fmtp,
-    },
-    errors::RtpError,
-    header::RtpHeader,
-    packet::sequencer::{RtpBufferAudioItem, RtpBufferItem, RtpBufferedSequencer},
-};
-
-use super::RtpMpeg4GenericPacket;
 
 pub mod de_interleaving;
 pub mod fragments;
